Finished
You can now use custom reactions for specified users, remove users from the list, clear the list, and basically everything I hoped for out of this project.

+++ auto.js
@@ -43,26 +43,16 @@ const prefix = config.prefix;
 client.on('message', message => {
     const args = message.content.slice(prefix.length).trim().split(/ +/g);
     const command = args.shift().toLowerCase();
-
-    // TODO: Add an else for this that catches if false
-    if (!!userList)
-    userList.forEach(function (userId) {
-        if (message.author.id == userId) {
-            // IDEA: Have way to custom select emoji here, or do they have
-            // to manually add it?
-            // IDEA: Could make each user have their own reaction array with sub-arrays
-            message.react("673643185997873182")
-            .then(console.log("Reacted to user!"))
-            .catch(function(err) {
-                console.error("[ERROR] Problem reacting! Message was probably deleted, but here's the full output anyways =>", err);
-            });
-        }
-    });
+    reactIfUserOnList(message);
 
     if (command == "add") {
-        var mentionedUser = getMentionedUser(message);
+        if (args.length == 1) {
+            console.log(`${prefix}add doesn't work that way! You need to include at least one reaction.`);
+            return;
+        }
 
         // Deletes the command message
+        var mentionedUser = getMentionedUser(message);
         message.member.lastMessage.delete().catch(console.error);
 
         if (!!mentionedUser) {
@@ -75,9 +65,13 @@ client.on('message', message => {
      
                     if (index < 0) {
                         currentUserList = json.users;
-                        currentUserList.push(mentionedUser);
+                        var newUser = {
+                            user: mentionedUser,
+                            reactions: generateReactionArray(args)
+                        }
 
-                        // To make it wait until the data has been retrieved
+                        // Add the user and their reaction array to the currentUserList
+                        currentUserList.push(newUser);
                         updateUserList(currentUserList);
                     } else {
                         console.log("That user is already on the reaction list!");
@@ -89,20 +83,21 @@ client.on('message', message => {
             console.log("Specify a user to add to the reaction list!");
         }
     } else if (command == "remove") {
-        var mentionedUser = getMentionedUser(message);
-
         // Deletes the command message
+        var mentionedUser = getMentionedUser(message);
         message.member.lastMessage.delete().catch(console.error);
 
         if (!!mentionedUser) {
-            var index = userList.indexOf(mentionedUser);
-     
-            if (index > -1) {
-                userList.splice(index, 1);
-                updateUserList(userList);
-            } else {
-                console.log("That user is not on the reaction list!");
-            }
+            var userNotFound = true;
+            userList.forEach(function (user, index) {
+                if (user.user == mentionedUser) {
+                    userList.splice(index, 1);
+                    updateUserList(userList);
+                    userNotFound = false;
+                }
+            });
+
+            if (userNotFound) console.log("That user is not on the reaction list!");
         } else {
             // Last message sent; user not specified
             console.log("Specify a user to remove from the reaction list!");
@@ -113,6 +108,44 @@ client.on('message', message => {
         message.member.lastMessage.delete().catch(console.error)
     }
 });
+
+/**
+ * Goes over the user list to see if the author is on it, and if
+ * they are it reacts the reaction array in order to their message.
+ * @param {object} message - the sent message object
+ */
+function reactIfUserOnList(message) {
+    if (!!userList) {
+        // Iterate over user list to see if user is on it
+        userList.forEach(async function (userData) {
+            var userId = userData.user;
+            if (message.author.id == userId) {
+                // Copied almost entirely from my multiple-reactions repository
+                var reactions = userData.reactions;
+                var breakLoop = false;
+
+                // Iterate over reaction array to add all reactions in order
+                for (i = 0; i < reactions.length; i++) {
+                    if (breakLoop) break;
+                    var userReaction = client.emojis.get(reactions[i]);
+                    
+                    // Discord native emoji or server/Nitro
+                    if (reactions[i].length < 8 || userReaction) {
+                        // https://discordjs.guide/popular-topics/reactions.html#reacting-in-order
+                        await message.react(reactions[i]).catch(function(err) {
+                            console.log("The last message that user sent was deleted! Cannot react...");
+                            breakLoop = true;
+                        });
+                    } else {
+                        console.log("Emoji not available on server or your client..."); 
+                    }
+                }
+            }
+        });
+    } else {
+        console.log("There is no user list! Did you delete a file while the program was running or something?");
+    }
+}
 
 // Either returns the mentioned user if there is one or null
 // if there wasn't a user mentioned.
@@ -145,6 +178,37 @@ function updateUserList(newUserArray) {
             userList = newUserArray;
         }
     });
+}
+
+
+/**
+ * Iterates over arguments passed and turns that into a list of
+ * reactions that will be sent when the command is called.
+ * @param {array} args - array of arguments passed from user
+ * @returns {array} array of reaction emojis
+ */
+function generateReactionArray(args) {
+    // Get reactions from arguments after `/set {commandName}`
+    var reactionArray = [];
+    for (i = 1; i < args.length; i++) {
+        // Gets the ID, not just the emoji
+        var emojiCode = "\\" + args[i];
+        
+        if (emojiCode.indexOf(':') != -1) {
+            // Gets the emoji ID of custom emojis
+            console.log("Gathering custom emoji ID...");
+            emojiCode = emojiCode.split(':')[2];
+            emojiCode = emojiCode.substring(0, emojiCode.length - 1);
+        } else {
+            // Remove the backslashes
+            emojiCode = args[i];
+        }
+
+        // [i-1] to ignore the commandName and start at 0 for reactions
+        reactionArray[i-1] = emojiCode;
+    }
+
+    return reactionArray;
 }
 
 client.login(config.botToken);
